optimise python script

+++ mpi-tests/test_nearest_algorithm.py
@@ -126,78 +126,99 @@ try:
     modulo = 100
 
     i = 0
-    for n_rank in args.n_ranks:
-        print("Running now jobs with", n_rank, " MPI ranks.", flush=True)
-        for seed in args.seeds:
-            rng = np.random.default_rng(seed)
-            for n in range(args.n_samples):
-                i = i + 1
+    for seed in args.seeds:
+        print("Running now jobs with seed", seed, flush=True)
+        rng = np.random.default_rng(seed)
+        for n in range(args.n_samples):
+            # -------------------------------------------------------------
+            # Set up the parcel configuration:
+            x_position = rng.uniform(low=0.0, high=1.0, size=n_parcels)
+            y_position = rng.uniform(low=0.0, high=1.0, size=n_parcels)
+            z_position = rng.uniform(low=0.0, high=1.0, size=n_parcels)
 
-                # -------------------------------------------------------------
-                # Set up the parcel configuration:
-                x_position = rng.uniform(low=0.0, high=1.0, size=n_parcels)
-                y_position = rng.uniform(low=0.0, high=1.0, size=n_parcels)
-                z_position = rng.uniform(low=0.0, high=1.0, size=n_parcels)
+            x_vorticity = rng.uniform(low=-10.0, high=10.0, size=n_parcels)
+            y_vorticity = rng.uniform(low=-10.0, high=10.0, size=n_parcels)
+            z_vorticity = rng.uniform(low=-10.0, high=10.0, size=n_parcels)
 
-                x_vorticity = rng.uniform(low=-10.0, high=10.0, size=n_parcels)
-                y_vorticity = rng.uniform(low=-10.0, high=10.0, size=n_parcels)
-                z_vorticity = rng.uniform(low=-10.0, high=10.0, size=n_parcels)
+            buoyancy = rng.uniform(low=-1.0, high=1.0, size=n_parcels)
 
-                buoyancy = rng.uniform(low=-1.0, high=1.0, size=n_parcels)
+            volume = rng.uniform(low=0.5 * vmin, high=1.5 * vmin, size=n_parcels)
 
-                volume = rng.uniform(low=0.5 * vmin, high=1.5 * vmin, size=n_parcels)
+            # lam = a / c
+            lam = rng.uniform(low=1.0, high=4.0, size=n_parcels)
 
-                # lam = a / c
-                lam = rng.uniform(low=1.0, high=4.0, size=n_parcels)
+            # lam2 = a / b
+            lam2 = rng.uniform(low=1.0, high=4.0, size=n_parcels)
 
-                # lam2 = a / b
-                lam2 = rng.uniform(low=1.0, high=4.0, size=n_parcels)
+            abc = 0.75 * volume / np.pi
 
-                abc = 0.75 * volume / np.pi
+            a2 = (abc * lam * lam2)  ** (2.0/3.0)
+            b2 = a2 / lam2 ** 2
+            c2 = a2 / lam ** 2
 
-                a2 = (abc * lam * lam2)  ** (2.0/3.0)
-                b2 = a2 / lam2 ** 2
-                c2 = a2 / lam ** 2
+            theta = rng.uniform(low=0.0, high=2.0*np.pi, size=n_parcels)
+            phi = rng.uniform(low=0.0, high=2.0*np.pi, size=n_parcels)
 
-                theta = rng.uniform(low=0.0, high=2.0*np.pi, size=n_parcels)
-                phi = rng.uniform(low=0.0, high=2.0*np.pi, size=n_parcels)
+            st = np.sin(theta)
+            ct = np.cos(theta)
+            sp = np.sin(phi)
+            cp = np.cos(phi)
 
-                st = np.sin(theta)
-                ct = np.cos(theta)
-                sp = np.sin(phi)
-                cp = np.cos(phi)
+            B11 = a2 * ct ** 2 * sp ** 2 + b2 * st ** 2 + c2 * ct ** 2 * cp ** 2
+            B12 = a2 * st * ct * sp ** 2 - b2 * st * ct + c2 * st * ct * cp ** 2
+            B13 = (a2 - c2) * ct * sp * cp
+            B22 = a2 * st ** 2 * sp ** 2 + b2 * ct ** 2 + c2 * st ** 2 * cp ** 2
+            B23 = (a2 - c2) * st * sp * cp
+            
+            # -------------------------------------------------------------
+            # Write the initial NetCDF file:
+            ncp.open('initial_parcels.nc')
 
-                B11 = a2 * ct ** 2 * sp ** 2 + b2 * st ** 2 + c2 * ct ** 2 * cp ** 2
-                B12 = a2 * st * ct * sp ** 2 - b2 * st * ct + c2 * st * ct * cp ** 2
-                B13 = (a2 - c2) * ct * sp * cp
-                B22 = a2 * st ** 2 * sp ** 2 + b2 * ct ** 2 + c2 * st ** 2 * cp ** 2
-                B23 = (a2 - c2) * st * sp * cp
+            ncp.add_box([0.0, 0.0, 0.0], [1.0, 1.0, 1.0], [nx, ny, nz])
 
-                # -------------------------------------------------------------
-                # Write the initial NetCDF file:
-                ncp.open('initial_parcels.nc')
+            ncp.add_dataset('x_position', x_position, unit='m')
+            ncp.add_dataset('y_position', y_position, unit='m')
+            ncp.add_dataset('z_position', z_position, unit='m')
 
-                ncp.add_box([0.0, 0.0, 0.0], [1.0, 1.0, 1.0], [nx, ny, nz])
+            ncp.add_dataset('buoyancy', buoyancy, unit='m/s^2')
 
-                ncp.add_dataset('x_position', x_position, unit='m')
-                ncp.add_dataset('y_position', y_position, unit='m')
-                ncp.add_dataset('z_position', z_position, unit='m')
+            ncp.add_dataset('volume', volume, unit='m^3')
 
-                ncp.add_dataset('buoyancy', buoyancy, unit='m/s^2')
+            ncp.add_dataset('x_vorticity', x_vorticity, unit='1/s')
+            ncp.add_dataset('y_vorticity', y_vorticity, unit='1/s')
+            ncp.add_dataset('z_vorticity', z_vorticity, unit='1/s')
 
-                ncp.add_dataset('volume', volume, unit='m^3')
+            ncp.add_dataset('B11', B11, unit='m^2')
+            ncp.add_dataset('B12', B12, unit='m^2')
+            ncp.add_dataset('B13', B13, unit='m^2')
+            ncp.add_dataset('B22', B22, unit='m^2')
+            ncp.add_dataset('B23', B23, unit='m^2')
 
-                ncp.add_dataset('x_vorticity', x_vorticity, unit='1/s')
-                ncp.add_dataset('y_vorticity', y_vorticity, unit='1/s')
-                ncp.add_dataset('z_vorticity', z_vorticity, unit='1/s')
+            ncp.close()
 
-                ncp.add_dataset('B11', B11, unit='m^2')
-                ncp.add_dataset('B12', B12, unit='m^2')
-                ncp.add_dataset('B13', B13, unit='m^2')
-                ncp.add_dataset('B22', B22, unit='m^2')
-                ncp.add_dataset('B23', B23, unit='m^2')
+            i = i + 1
 
-                ncp.close()
+            try:
+                cmd = 'mpirun -np 1 '
+                if args.cmd == 'srun':
+                    cmd = 'srun --nodes=1 --ntasks=1 --ntasks-per-node=1 --exact '
+                subprocess.run(args=cmd + exec_serial + flags,
+                               shell=True,
+                               check=True,
+                               timeout=120,
+                               stdout=subprocess.DEVNULL,
+                               stderr=subprocess.STDOUT)
+            except subprocess.CalledProcessError:
+                print('Error in running the serial version.', flush=True)
+
+            if os.path.exists('serial_final_0000000001_parcels.nc'):
+                ncrs.open('serial_final_0000000001_parcels.nc')
+                n_merges = n_merges + n_parcels - ncrs.get_num_parcels(step=0)
+                ncrs.close()
+                print ("Sample", i, "generated.", flush=True)
+
+            for n_rank in args.n_ranks:
+                #print("Run sample with", n_rank, "MPI ranks", flush=True)
 
                 # -------------------------------------------------------------
                 # Run the serial and parallel versions of the nearest + merging algorithm:
@@ -220,20 +241,6 @@ try:
                     print('Error in running the parallel version.', flush=True)
                     failed = True
 
-                try:
-                    cmd = 'mpirun -np 1 '
-                    if args.cmd == 'srun':
-                        cmd = 'srun --nodes=1 --ntasks=1 --ntasks-per-node=1 --exact '
-                    subprocess.run(args=cmd + exec_serial + flags,
-                                   shell=True,
-                                   check=True,
-                                   timeout=120,
-                                   stdout=subprocess.DEVNULL,
-                                   stderr=subprocess.STDOUT)
-                except subprocess.CalledProcessError:
-                    print('Error in running the serial version.', flush=True)
-                    failed = True
-
                 if not os.path.exists('serial_final_0000000001_parcels.nc') and \
                    not os.path.exists('parallel_final_0000000001_parcels.nc'):
                     failed = True
@@ -251,8 +258,6 @@ try:
                     ind2 = np.lexsort((ncrp.get_dataset(step=0, name='x_position'),
                                     ncrp.get_dataset(step=0, name='y_position'),
                                     ncrp.get_dataset(step=0, name='z_position')))
-
-                    n_merges = n_merges + n_parcels - ncrs.get_num_parcels(step=0)
 
                     failed = (not ncrs.get_num_parcels(step=0) == ncrp.get_num_parcels(step=0))
 
@@ -277,20 +282,21 @@ try:
                         os.rename('initial_parcels.nc',
                                   'initial_' + n_str + '_parcels.nc')
                     if os.path.exists('serial_final_0000000001_parcels.nc'):
-                        os.rename('serial_final_0000000001_parcels.nc',
-                                  'serial_fail_' + n_str + '_parcels.nc')
+                        os.system('cp serial_final_0000000001_parcels.nc serial_fail_' + n_str + '_parcels.nc')  
                     if os.path.exists('parallel_final_0000000001_parcels.nc'):
                         os.rename('parallel_final_0000000001_parcels.nc',
                                   'parallel_fail_' + n_str + '_parcels.nc')
                 else:
-                    os.remove('initial_parcels.nc')
-                    os.remove('serial_final_0000000001_parcels.nc')
                     os.remove('parallel_final_0000000001_parcels.nc')
 
-                # -------------------------------------------------------------
-                # Intermediate info:
-                if args.verbose and (i % modulo == 0):
-                    print("#samples, #fails, #merges: ", i, n_fails, n_merges, flush=True)
+            # -------------------------------------------------------------
+            # Intermediate info:
+            if args.verbose and (i % modulo == 0):
+                print("#samples, #fails, #merges: ", i, n_fails, n_merges, flush=True)
+
+            os.remove('initial_parcels.nc')
+            if os.path.exists('serial_final_0000000001_parcels.nc'):
+                os.remove('serial_final_0000000001_parcels.nc')
 
     # -------------------------------------------------------------------------
     # Print summary:
